Move 'FeedImageMapper' to it's own file
